Lesson4.level1 final commit

+++ src/student_yevgeniy_tolks/lesson_4/level_1/Task_4.java
@@ -6,7 +6,7 @@ class Task_4 {
     public static void main(String[] args) {
         Scanner scanner=new Scanner(System.in);
         System.out.println("Enter any number: ");
-        double number = scanner.nextDouble();
+        int number = scanner.nextInt();
 
         if(number%2==0){
             System.out.println(number + " is even number");
